fw: move vector table section into every soc file
If any SoC file adds a new section to flash, it will get the first
offset in the flash, triggering the ISR vector section assert. None
of the existing SoCs needed to add new sections, so it was ok to keep
the entry in the common file. However, SF32LB will need to add an extra
entry that triggers this problem. If we ever preprocess the ld files,
or generate them in a more advanced manner, we could improve this, but
now I do not see any other immediate solution (maybe I'm just tired).

Signed-off-by: Gerard Marull-Paretas <gerard@teslabs.com>